Progress calculations, action response handling
Progress now excludes the last step in, since in order to complete a
step you must go to the next step. So a "three" step wizard would
actually show "0/2" instead of "0/3" because you only need to complete
two steps in order to get to the last step of the wizard.

Action responses can now return false to prevent going anywhere without
throwing an error about unknown states. Usually this indicates that
something was not present in order to calculate where to go next, so the
user should handle this case (such as alerting the user).

Also added the $.isJquery method since the code for it was duplicated in
a few places.