_fastForward() rewrite, default branch bug fix
The _fastForward() method now takes two parameters: fromIndex and
toIndex. There is now a better indication as to where the fast forward
failed which is indicated in the error that is thrown.

Fixed a bug in that stepsWrapper was incorrectly being assigned to the
parent of every step, instead of just the first step, which was causing
the cached branches to be invalid.

Also in this commit were minor modifications to thrown messages and some
structural organization.