Fix inability to select initialStep with history
Always select first step with initialStep as array of arguments (apply).
This allows the user to specify any arguments that would normally be
allowed in the select method.

Make the argument 'relative' optional in the select method. This allows
the user to pass only step and history without entering intermediate
arguments.